+++ apps/toolkit-app/app/auth/mutations/forgotPassword.ts
@@ -1,6 +1,6 @@
 import { generateToken, hash256 } from "@blitzjs/auth"
 import { resolver } from "@blitzjs/rpc"
-import { db } from "db"
+import db from "db"
 import { forgotPasswordMailer } from "mailers/forgotPasswordMailer"
 import { ForgotPassword } from "../validations"