+++ apps/toolkit-app/app/blitz-server.ts
@@ -1,14 +1,13 @@
 import { setupBlitzServer } from "@blitzjs/next"
 import { AuthServerPlugin, PrismaStorage } from "@blitzjs/auth"
-import { db } from "db"
+import db from "db"
 import { simpleRolesIsAuthorized } from "@blitzjs/auth"
 
 const { gSSP, gSP, api } = setupBlitzServer({
   plugins: [
     AuthServerPlugin({
       cookiePrefix: "web-cookie-prefix",
-      // TODO fix type
-      storage: PrismaStorage(db as any),
+      storage: PrismaStorage(db),
       isAuthorized: simpleRolesIsAuthorized,
     }),
   ],