+++ integration-tests/middleware/app/blitz-client.ts
@@ -1,0 +1,5 @@
+import {setupBlitzClient} from "@blitzjs/next"
+
+const {withBlitz} = setupBlitzClient({plugins: []})
+
+export {withBlitz}