+++ integration-tests/rpc/next.config.js
@@ -1,4 +1,4 @@
 const {withBlitz} = require("@blitzjs/next")
 module.exports = withBlitz({
-  target: 'experimental-serverless-trace',
+  target: "experimental-serverless-trace",
 })