+++ integration-tests/trailing-slash/db/seed.ts
@@ -1,4 +1,4 @@
-import {prisma} from "./index"
+import prisma from "./index"
 
 const seed = async () => {
   await prisma.$reset()