+++ packages/blitz-next/src/index-server.ts
@@ -17,12 +17,19 @@ import type {
   MiddlewareResponse,
 } from "blitz"
 import {handleRequestWithMiddleware, startWatcher, stopWatcher} from "blitz"
-import {dehydrate, getQueryKey, getInfiniteQueryKey, loaderClient, loaderServer} from "@blitzjs/rpc"
+import {
+  dehydrate,
+  getInfiniteQueryKey,
+  getQueryKey,
+  installWebpackConfig,
+  InstallWebpackConfigOptions,
+  ResolverPathOptions,
+} from "@blitzjs/rpc"
 import {DefaultOptions, QueryClient} from "react-query"
 import {IncomingMessage, ServerResponse} from "http"
 import {withSuperJsonProps} from "./superjson"
-import {ResolverBasePath} from "@blitzjs/rpc/src/index-server"
 import {ParsedUrlQuery} from "querystring"
+import {PreviewData} from "next/types"
 
 export * from "./index-browser"
 
@@ -40,21 +47,32 @@ export type NextApiHandler = (
 
 type SetupBlitzOptions = {
   plugins: BlitzServerPlugin<RequestMiddleware, Ctx>[]
+  onError?: (err: Error) => void
 }
 
-export type BlitzGSSPHandler<TProps, Query extends ParsedUrlQuery = ParsedUrlQuery> = ({
+export type BlitzGSSPHandler<
+  TProps,
+  Query extends ParsedUrlQuery = ParsedUrlQuery,
+  PD extends PreviewData = PreviewData,
+> = ({
   ctx,
   req,
   res,
   ...args
-}: Parameters<GetServerSideProps<TProps>>[0] & {ctx: Ctx}) => ReturnType<
-  GetServerSideProps<TProps, Query>
+}: Parameters<GetServerSideProps<TProps, Query, PD>>[0] & {ctx: Ctx}) => ReturnType<
+  GetServerSideProps<TProps, Query, PD>
 >
 
-export type BlitzGSPHandler<TProps> = ({
+export type BlitzGSPHandler<
+  TProps,
+  Query extends ParsedUrlQuery = ParsedUrlQuery,
+  PD extends PreviewData = PreviewData,
+> = ({
   ctx,
   ...args
-}: Parameters<GetStaticProps<TProps>>[0] & {ctx: Ctx}) => ReturnType<GetStaticProps<TProps>>
+}: Parameters<GetStaticProps<TProps, Query, PD>>[0] & {ctx: Ctx}) => ReturnType<
+  GetStaticProps<TProps, Query, PD>
+>
 
 export type BlitzAPIHandler = (
   req: Parameters<NextApiHandler>[0],
@@ -62,14 +80,14 @@ export type BlitzAPIHandler = (
   ctx: Ctx,
 ) => ReturnType<NextApiHandler>
 
-export const setupBlitzServer = ({plugins}: SetupBlitzOptions) => {
+export const setupBlitzServer = ({plugins, onError}: SetupBlitzOptions) => {
   const middlewares = plugins.flatMap((p) => p.requestMiddlewares)
   const contextMiddleware = plugins.flatMap((p) => p.contextMiddleware).filter(Boolean)
 
   const gSSP =
-    <TProps, Query extends ParsedUrlQuery = ParsedUrlQuery>(
-      handler: BlitzGSSPHandler<TProps, Query>,
-    ): GetServerSideProps<TProps> =>
+    <TProps, Query extends ParsedUrlQuery = ParsedUrlQuery, PD extends PreviewData = PreviewData>(
+      handler: BlitzGSSPHandler<TProps, Query, PD>,
+    ): GetServerSideProps<TProps, Query, PD> =>
     async ({req, res, ...rest}) => {
       await handleRequestWithMiddleware<IncomingMessage, ServerResponse>(req, res, middlewares)
       const ctx = contextMiddleware.reduceRight(
@@ -93,12 +111,19 @@ export const setupBlitzServer = ({plugins}: SetupBlitzOptions) => {
       ctx.prefetchQuery = prefetchQuery
       ctx.prefetchInfiniteQuery = (...args) => prefetchQuery(...args, true)
 
-      const result = await handler({req, res, ctx, ...rest})
-      return withSuperJsonProps(withDehydratedState(result, queryClient))
+      try {
+        const result = await handler({req, res, ctx, ...rest})
+        return withSuperJsonProps(withDehydratedState(result, queryClient))
+      } catch (err: any) {
+        onError?.(err)
+        throw err
+      }
     }
 
   const gSP =
-    <TProps>(handler: BlitzGSPHandler<TProps>): GetStaticProps<TProps> =>
+    <TProps, Query extends ParsedUrlQuery = ParsedUrlQuery, PD extends PreviewData = PreviewData>(
+      handler: BlitzGSPHandler<TProps, Query, PD>,
+    ): GetStaticProps<TProps, Query, PD> =>
     async (context) => {
       const ctx = contextMiddleware.reduceRight((y, f) => (f ? f(y) : y), {} as Ctx)
       let queryClient: null | QueryClient = null
@@ -118,8 +143,13 @@ export const setupBlitzServer = ({plugins}: SetupBlitzOptions) => {
       ctx.prefetchQuery = prefetchQuery
       ctx.prefetchInfiniteQuery = (...args) => prefetchQuery(...args, true)
 
-      const result = await handler({...context, ctx: ctx})
-      return withSuperJsonProps(withDehydratedState(result, queryClient))
+      try {
+        const result = await handler({...context, ctx: ctx})
+        return withSuperJsonProps(withDehydratedState(result, queryClient))
+      } catch (err: any) {
+        onError?.(err)
+        throw err
+      }
     }
 
   const api =
@@ -128,7 +158,8 @@ export const setupBlitzServer = ({plugins}: SetupBlitzOptions) => {
       try {
         await handleRequestWithMiddleware(req, res, middlewares)
         return handler(req, res, res.blitzCtx)
-      } catch (error) {
+      } catch (error: any) {
+        onError?.(error)
         return res.status(400).send(error)
       }
     }
@@ -138,57 +169,11 @@ export const setupBlitzServer = ({plugins}: SetupBlitzOptions) => {
 
 export interface BlitzConfig extends NextConfig {
   blitz?: {
-    resolverBasePath?: ResolverBasePath
+    resolverPath?: ResolverPathOptions
     customServer?: {
       hotReload?: boolean
     }
   }
-}
-
-interface WebpackRuleOptions {
-  resolverBasePath?: ResolverBasePath
-}
-
-interface WebpackRule {
-  test: RegExp
-  use: Array<{
-    loader: string
-    options: WebpackRuleOptions
-  }>
-}
-
-interface InstallWebpackConfigOptions {
-  webpackConfig: {
-    module: {
-      rules: WebpackRule[]
-    }
-  }
-  nextConfig: BlitzConfig
-}
-
-export function installWebpackConfig({webpackConfig, nextConfig}: InstallWebpackConfigOptions) {
-  const options: WebpackRuleOptions = {
-    resolverBasePath: nextConfig.blitz?.resolverBasePath,
-  }
-
-  webpackConfig.module.rules.push({
-    test: /\/\[\[\.\.\.blitz]]\.[jt]s$/,
-    use: [
-      {
-        loader: loaderServer,
-        options,
-      },
-    ],
-  })
-  webpackConfig.module.rules.push({
-    test: /[\\/](queries|mutations)[\\/]/,
-    use: [
-      {
-        loader: loaderClient,
-        options,
-      },
-    ],
-  })
 }
 
 export function withBlitz(nextConfig: BlitzConfig = {}) {
@@ -211,7 +196,12 @@ export function withBlitz(nextConfig: BlitzConfig = {}) {
 
   const config = Object.assign({}, nextConfig, {
     webpack: (config: InstallWebpackConfigOptions["webpackConfig"], options: any) => {
-      installWebpackConfig({webpackConfig: config, nextConfig})
+      installWebpackConfig({
+        webpackConfig: config,
+        webpackRuleOptions: {
+          resolverPath: nextConfig.blitz?.resolverPath,
+        },
+      })
       if (typeof nextConfig.webpack === "function") {
         return nextConfig.webpack(config, options)
       }