+++ packages/blitz-rpc/src/data-client/rpc.ts
@@ -1,5 +1,5 @@
 import {normalizePathTrailingSlash} from "next/dist/client/normalize-trailing-slash"
-import {addBasePath} from "next/dist/shared/lib/router/router"
+import {addBasePath} from "next/dist/client/add-base-path"
 import {deserialize, serialize} from "superjson"
 import {SuperJSONResult} from "superjson/dist/types"
 import {CSRFTokenMismatchError, isServer} from "blitz"