+++ packages/blitz-rpc/src/index-server.ts
@@ -1,7 +1,6 @@
-import {assert, baseLogger, Ctx, newLine, prettyMs} from "blitz"
+import {assert, Ctx, prettyMs} from "blitz"
 import {NextApiRequest, NextApiResponse} from "next"
 import {deserialize, serialize as superjsonSerialize} from "superjson"
-import chalk from "chalk"
 import {resolve} from "path"
 
 // TODO - optimize end user server bundles by not exporting all client stuff here
@@ -13,6 +12,7 @@ export * from "./resolver"
 function isObject(value: unknown): value is Record<string | symbol, unknown> {
   return typeof value === "object" && value !== null
 }
+
 function getGlobalObject<T extends Record<string, unknown>>(key: string, defaultValue: T): T {
   assert(key.startsWith("__internal_blitz"), "unsupported key")
   if (typeof global === "undefined") {
@@ -25,7 +25,7 @@ function getGlobalObject<T extends Record<string, unknown>>(key: string, default
 
 type Resolver = (...args: unknown[]) => Promise<unknown>
 type ResolverFiles = Record<string, () => Promise<{default?: Resolver}>>
-export type ResolverBasePath = "queries|mutations" | "root" | undefined
+export type ResolverPathOptions = "queries|mutations" | "root" | ((path: string) => string)
 
 // We define `global.__internal_blitzRpcResolverFiles` to ensure we use the same global object.
 // Needed for Next.js. I'm guessing that Next.js is including the `node_modules/` files in a seperate bundle than user files.
@@ -50,8 +50,53 @@ export function __internal_addBlitzRpcResolver(
 }
 
 const dir = __dirname + (() => "")() // trick to avoid `@vercel/ncc` to glob import
-export const loaderServer = resolve(dir, "./loader-server.cjs")
-export const loaderClient = resolve(dir, "./loader-client.cjs")
+const loaderServer = resolve(dir, "./loader-server.cjs")
+const loaderClient = resolve(dir, "./loader-client.cjs")
+
+interface WebpackRuleOptions {
+  resolverPath: ResolverPathOptions | undefined
+}
+
+interface WebpackRule {
+  test: RegExp
+  use: Array<{
+    loader: string
+    options: WebpackRuleOptions
+  }>
+}
+
+export interface InstallWebpackConfigOptions {
+  webpackConfig: {
+    module: {
+      rules: WebpackRule[]
+    }
+  }
+  webpackRuleOptions: WebpackRuleOptions
+}
+
+export function installWebpackConfig({
+  webpackConfig,
+  webpackRuleOptions,
+}: InstallWebpackConfigOptions) {
+  webpackConfig.module.rules.push({
+    test: /\/\[\[\.\.\.blitz]]\.[jt]s$/,
+    use: [
+      {
+        loader: loaderServer,
+        options: webpackRuleOptions,
+      },
+    ],
+  })
+  webpackConfig.module.rules.push({
+    test: /[\\/](queries|mutations)[\\/]/,
+    use: [
+      {
+        loader: loaderClient,
+        options: webpackRuleOptions,
+      },
+    ],
+  })
+}
 
 // ----------
 // END LOADER
@@ -109,14 +154,6 @@ export function rpcHandler(config: RpcConfig) {
       throw new Error("No default export for resolver path: " + routePath)
     }
 
-    const log = baseLogger().getChildLogger({
-      prefix: [relativeRoutePath + "()"],
-    })
-
-    const customChalk = new chalk.Instance({
-      level: log.settings.type === "json" ? 0 : chalk.level,
-    })
-
     if (req.method === "HEAD") {
       // We used to initiate database connection here
       res.status(200).end()
@@ -126,7 +163,7 @@ export function rpcHandler(config: RpcConfig) {
 
       if (typeof req.body.params === "undefined") {
         const error = {message: "Request body is missing the `params` key"}
-        log.error(error.message)
+        console.error(error.message)
         res.status(400).json({
           result: null,
           error,
@@ -140,11 +177,11 @@ export function rpcHandler(config: RpcConfig) {
           meta: req.body.meta?.params,
         })
 
-        log.info(customChalk.dim("Starting with input:"), data ? data : JSON.stringify(data))
+        console.info("Starting with input:", data ? data : JSON.stringify(data))
         const startTime = Date.now()
         const result = await resolver(data, (res as any).blitzCtx)
         const resolverDuration = Date.now() - startTime
-        log.debug(customChalk.dim("Result:"), result ? result : JSON.stringify(result))
+        console.debug("Result:", result ? result : JSON.stringify(result))
 
         const serializerStartTime = Date.now()
         const serializedResult = superjsonSerialize(result)
@@ -158,30 +195,27 @@ export function rpcHandler(config: RpcConfig) {
             result: serializedResult.meta,
           },
         })
-        log.debug(
-          customChalk.dim(
-            `Next.js serialization:${prettyMs(Date.now() - nextSerializerStartTime)}`,
-          ),
-        )
+        console.debug(`Next.js serialization:${prettyMs(Date.now() - nextSerializerStartTime)}`)
         const serializerDuration = Date.now() - serializerStartTime
         const duration = Date.now() - startTime
 
-        log.info(
-          customChalk.dim(
-            `Finished: resolver:${prettyMs(resolverDuration)} serializer:${prettyMs(
-              serializerDuration,
-            )} total:${prettyMs(duration)}`,
-          ),
+        console.info(
+          `Finished: resolver:${prettyMs(resolverDuration)} serializer:${prettyMs(
+            serializerDuration,
+          )} total:${prettyMs(duration)}`,
         )
-        newLine()
+        console.log("\n")
 
         return
       } catch (error: any) {
         if (error._clearStack) {
           delete error.stack
         }
-        log.error(error)
-        newLine()
+
+        config.onError?.(error)
+
+        console.error(error)
+        console.log("\n")
 
         if (!error.statusCode) {
           error.statusCode = 500
@@ -200,7 +234,7 @@ export function rpcHandler(config: RpcConfig) {
       }
     } else {
       // Everything else is error
-      log.warn(`${req.method} method not supported`)
+      console.warn(`${req.method} method not supported`)
       res.status(404).end()
       return
     }