+++ packages/blitz-auth/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @blitzjs/auth
 
+## 2.0.0-alpha.53
+
+### Patch Changes
+
+- blitz@2.0.0-alpha.53
+
+## 2.0.0-alpha.52
+
+### Patch Changes
+
+- blitz@2.0.0-alpha.52
+
+## 2.0.0-alpha.51
+
+### Patch Changes
+
+- blitz@2.0.0-alpha.51
+
+## 2.0.0-alpha.50
+
+### Patch Changes
+
+- blitz@2.0.0-alpha.50
+
+## 2.0.0-alpha.49
+
+### Patch Changes
+
+- blitz@2.0.0-alpha.49
+
+## 2.0.0-alpha.48
+
+### Patch Changes
+
+- Updated dependencies [93851d90]
+  - blitz@2.0.0-alpha.48
+
+## 2.0.0-alpha.47
+
+### Patch Changes
+
+- 1d9804a6: Remove references to the logging package
+- Updated dependencies [1d9804a6]
+  - blitz@2.0.0-alpha.47
+
+## 2.0.0-alpha.46
+
+### Patch Changes
+
+- blitz@2.0.0-alpha.46
+
 ## 2.0.0-alpha.45
 
 ### Patch Changes